textile-device-monitor/unigraco-web/requirements.txt
gradio==4.44.1
numpy==1.24.3
opencv-python-headless==4.8.1.78
torchvision==0.15.2
PyYAML==6.0
easydict==1.9
timm==0.6.11
mmcv==1.6.2
albumentations==0.5.2
Cython>=0.29.32
loralib==0.1.2
pycocotools
ftfy
regex
huggingface_hub==0.20.3
tensorboard==2.10.0
